working on floating point numbers

+++ src/main/java/rocks/zipcode/atm/login/LoggedInComponent.java
@@ -34,7 +34,7 @@ public class LoggedInComponent {
 
         Button btnDeposit = new Button("Deposit");
         btnDeposit.setOnAction(e -> {
-            int amount = Integer.parseInt(field.getText());
+            Float amount = Float.valueOf(Integer.parseInt(field.getText()));
             cashMachine.deposit(amount);
 
             areaInfo.setText(cashMachine.toString());
